requirements.txt
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.10
astroid @ file:///C:/ci/astroid_1592495992870/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
Authlib==0.15.2
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
Babel==2.8.0
backcall==0.2.0
bcrypt @ file:///C:/ci/bcrypt_1597936262193/work
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
boto3==1.16.47
botocore==1.19.47
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600681203656/work
chardet==3.0.4
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama==0.4.3
cryptography @ file:///C:/ci/cryptography_1598892109987/work
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-storages==1.11.1
docutils==0.16
entrypoints==0.3
flake8==3.8.3
future==0.18.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593431910144/work
intervaltree @ file:///tmp/build/80754af9/intervaltree_1598376443606/work
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils==0.2.0
isort @ file:///C:/ci/isort_1600683642740/work
jedi @ file:///C:/ci/jedi_1592841982010/work
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
keyring @ file:///C:/ci/keyring_1598884806717/work
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.19.3
numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
packaging==20.4
pandocfilters==1.4.2
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
psutil @ file:///C:/ci/psutil_1598352185190/work
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1598885001695/work
pyflakes==2.2.0
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pylint @ file:///C:/ci/pylint_1598624066230/work
PyMySQL==0.10.1
PyNaCl @ file:///C:/ci/pynacl_1595009245871/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1600278539111/work
python-language-server @ file:///tmp/build/80754af9/python-language-server_1600454544709/work
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rope==0.17.0
Rtree==0.9.4
s3transfer==0.3.3
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx @ file:///tmp/build/80754af9/sphinx_1597428793432/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder @ file:///C:/ci/spyder_1599057304706/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1599051725691/work
sqlparse==0.3.1
stripe==2.55.0
testpath==0.4.4
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson @ file:///C:/ci/ujson_1600709405842/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
watchdog @ file:///C:/ci/watchdog_1593446204230/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zipp==3.1.0
